Location permission quickfix
Can now handle the location service being disabled at
arbitrary times. Will re-subscribe if it is enabled.
Should now be impossible to reach a state where location
data is available but the app doesn't use it or doesn't
offer to ask for permission to use it.

Also show button to center on user unconditionally.